Refactor App by extracting PDF export and connection logic
